+++ packages/react/ds-global/src/lib/Badge/Badge.stories.tsx
@@ -1,5 +1,3 @@
-/* @canonical/generator-ds 0.10.0-experimental.2 */
-
 import { MODIFIER_FAMILIES } from "@canonical/ds-types";
 import type { Meta, StoryFn, StoryObj } from "@storybook/react-vite";
 import Component from "./Badge.js";