+++ packages/react/ds-global/src/lib/Card/common/Header/Header.tsx
@@ -1,5 +1,3 @@
-/* @canonical/generator-ds 0.10.0-experimental.2 */
-
 import type React from "react";
 import type { HeaderProps } from "./types.js";
 import "./styles.css";
@@ -10,7 +8,6 @@ const componentCssClassName = "ds card-header";
  * Header component for Card
  *
  * @implements ds:global.subcomponent.card-header
- * @returns {React.ReactElement} - Rendered Header
  */
 const Header = ({
   children,