+++ packages/react/ds-global/src/lib/Card/index.ts
@@ -1,4 +1,2 @@
-/* @canonical/generator-ds 0.10.0-experimental.2 */
-
-export { default } from "./Card.js";
+export { default as Card } from "./Card.js";
 export type * from "./types.js";