+++ packages/react/ds-global/src/lib/Chip/index.ts
@@ -1,3 +1,2 @@
-/* @canonical/generator-canonical-ds 0.4.0-experimental.0 */
 export { default as Chip } from "./Chip.js";
-export * from "./types.js";
+export type * from "./types.js";