+++ packages/react/ds-global/src/lib/Tooltip/common/TooltipArea/index.ts
@@ -1,3 +1,2 @@
-/* @canonical/generator-ds 0.9.0-experimental.4 */
 export { default as TooltipArea } from "./TooltipArea.js";
-export * from "./types.js";
+export type * from "./types.js";