+++ packages/react/ds-global/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.15.0](https://github.com/canonical/ds25/compare/v0.15.0-experimental.0...v0.15.0) (2026-02-20)
+
+
+### Bug Fixes
+
+* **react-ds-global:** Fixes transitional export patterns for components ([#426](https://github.com/canonical/ds25/issues/426)) ([db8a1db](https://github.com/canonical/ds25/commit/db8a1dba10419153f6be82ffee570c9db929dff7))
+
+
+
+
+
 # [0.15.0-experimental.0](https://github.com/canonical/ds25/compare/v0.14.0...v0.15.0-experimental.0) (2026-02-17)
 
 **Note:** Version bump only for package @canonical/react-ds-global