+++ packages/svelte/ds-app-launchpad/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.15.0](https://github.com/canonical/pragma/compare/v0.15.0-experimental.0...v0.15.0) (2026-02-20)
+
+**Note:** Version bump only for package @canonical/svelte-ds-app-launchpad
+
+
+
+
+
 # [0.15.0-experimental.0](https://github.com/canonical/pragma/compare/v0.14.0...v0.15.0-experimental.0) (2026-02-17)
 
 **Note:** Version bump only for package @canonical/svelte-ds-app-launchpad